Break out test claseses to their own files
